seperate descrptors and energy fitting from model
