+++ source/train/DescrptSeR.py
@@ -2,7 +2,6 @@ import os,warnings
 import numpy as np
 import tensorflow as tf
 from deepmd.common import j_must_have, j_must_have_d, j_have
-from deepmd.Model import Model
 
 from deepmd.RunOptions import global_tf_float_precision
 from deepmd.RunOptions import global_np_float_precision
@@ -14,7 +13,7 @@ module_path = os.path.dirname(os.path.realpath(__file__)) + "/"
 assert (os.path.isfile (module_path  + "libop_abi.so" )), "op module does not exist"
 op_module = tf.load_op_library(module_path + "libop_abi.so")
 
-class ModelSeR (Model):
+class DescrptSeR ():
     def __init__ (self, jdata):
         # descrpt config
         self.use_smooth = False
@@ -29,14 +28,6 @@ class ModelSeR (Model):
             self.rcut_smth = jdata['rcut_smth']
         else :
             self.rcut_smth = self.rcut
-        # fparam
-        self.numb_fparam = 0
-        if j_have(jdata, 'numb_fparam') :
-            self.numb_fparam = jdata['numb_fparam']
-	# type_map
-        self.type_map = []
-        if j_have(jdata, 'type_map') :
-            self.numb_fparam = jdata['type_map']
         # filter of smooth version
         if j_have(jdata, 'coord_norm') :
             self.coord_norm = jdata['coord_norm']
@@ -53,24 +44,6 @@ class ModelSeR (Model):
         self.ndescrpt_a = self.nnei_a * 4
         self.ndescrpt_r = self.nnei_r * 1
         self.ndescrpt = self.nnei_r
-        # network size
-        self.n_neuron = j_must_have_d (jdata, 'fitting_neuron', ['n_neuron'])
-        self.resnet_dt = True
-        if j_have(jdata, 'resnet_dt') :
-            warnings.warn("the key \"%s\" is deprecated, please use \"%s\" instead" % ('resnet_dt','fitting_resnet_dt'))
-            self.resnet_dt = jdata['resnet_dt']
-        if j_have(jdata, 'fitting_resnet_dt') :
-            self.resnet_dt = jdata['fitting_resnet_dt']
-        self.type_fitting_net = False            
-
-        # short-range tab
-        if 'use_srtab' in jdata :
-            self.srtab = TabInter(jdata['use_srtab'])
-            self.smin_alpha = j_must_have(jdata, 'smin_alpha')
-            self.sw_rmin = j_must_have(jdata, 'sw_rmin')
-            self.sw_rmax = j_must_have(jdata, 'sw_rmax')
-        else :
-            self.srtab = None
 
         self.seed = None
         if j_have (jdata, 'seed') :
@@ -83,11 +56,8 @@ class ModelSeR (Model):
     def get_ntypes (self) :
         return self.ntypes
 
-    def get_numb_fparam (self) :
-        return self.numb_fparam
-
-    def get_type_map (self) :
-        return self.type_map
+    def get_dim_out (self) :
+        return self.filter_neuron[-1]
 
     def compute_dstats (self,
                         data_coord, 
@@ -123,21 +93,17 @@ class ModelSeR (Model):
 
         return davg, dstd
 
-
-    def build_interaction (self, 
-                           coord_, 
-                           atype_,
-                           natoms,
-                           box, 
-                           mesh,
-                           fparam,
-                           davg = None, 
-                           dstd = None,
-                           bias_atom_e = None,
-                           suffix = '', 
-                           reuse_attr = None,
-                           reuse_weights = None):
-        with tf.variable_scope('model_attr' + suffix, reuse = reuse_attr) :
+    def build (self, 
+               coord_, 
+               atype_,
+               natoms,
+               box, 
+               mesh,
+               davg = None, 
+               dstd = None,
+               suffix = '', 
+               reuse = None):
+        with tf.variable_scope('model_attr' + suffix, reuse = reuse) :
             if davg is None:
                 davg = np.zeros([self.ntypes, self.ndescrpt]) 
             if dstd is None:
@@ -148,12 +114,6 @@ class ModelSeR (Model):
             t_ntypes = tf.constant(self.ntypes, 
                                    name = 'ntypes', 
                                    dtype = tf.int32)
-            t_dfparam = tf.constant(self.numb_fparam, 
-                                    name = 'dfparam', 
-                                    dtype = tf.int32)
-            t_tmap = tf.constant(' '.join(self.type_map), 
-                                 name = 'tmap', 
-                                 dtype = tf.string)
             self.t_avg = tf.get_variable('t_avg', 
                                          davg.shape, 
                                          dtype = global_tf_float_precision,
@@ -164,27 +124,15 @@ class ModelSeR (Model):
                                          dtype = global_tf_float_precision,
                                          trainable = False,
                                          initializer = tf.constant_initializer(dstd, dtype = global_tf_float_precision))
-            if self.srtab is not None :
-                tab_info, tab_data = self.srtab.get()
-                self.tab_info = tf.get_variable('t_tab_info',
-                                                tab_info.shape,
-                                                dtype = tf.float64,
-                                                trainable = False,
-                                                initializer = tf.constant_initializer(tab_info, dtype = tf.float64))
-                self.tab_data = tf.get_variable('t_tab_data',
-                                                tab_data.shape,
-                                                dtype = tf.float64,
-                                                trainable = False,
-                                                initializer = tf.constant_initializer(tab_data, dtype = tf.float64))
 
         coord = tf.reshape (coord_, [-1, natoms[1] * 3])
         atype = tf.reshape (atype_, [-1, natoms[1]])
 
-        descrpt, descrpt_deriv, rij, nlist \
+        self.descrpt, self.descrpt_deriv, self.rij, self.nlist \
             = op_module.descrpt_se_r (coord,
                                       atype,
                                       natoms,
-                                      box,                                    
+                                      box,
                                       mesh,
                                       self.t_avg,
                                       self.t_std,
@@ -192,144 +140,48 @@ class ModelSeR (Model):
                                       rcut_smth = self.rcut_smth,
                                       sel = self.sel_r)
 
-        descrpt_reshape = tf.reshape(descrpt, [-1, self.ndescrpt])
+        self.dout = self._pass_filter(self.descrpt, natoms, suffix = suffix, reuse = reuse)
 
-        atom_ener = self.build_atom_net (descrpt_reshape, 
-                                         fparam, 
-                                         natoms, 
-                                         bias_atom_e = bias_atom_e, 
-                                         reuse = reuse_weights, 
-                                         suffix = suffix)
+        return self.dout
 
-        if self.srtab is not None :
-            sw_lambda, sw_deriv \
-                = op_module.soft_min_switch(atype, 
-                                            rij, 
-                                            nlist,
-                                            natoms,
-                                            sel_a = self.sel_a,
-                                            sel_r = self.sel_r,
-                                            alpha = self.smin_alpha,
-                                            rmin = self.sw_rmin,
-                                            rmax = self.sw_rmax)            
-            inv_sw_lambda = 1.0 - sw_lambda
-            # NOTICE:
-            # atom energy is not scaled, 
-            # force and virial are scaled
-            tab_atom_ener, tab_force, tab_atom_virial \
-                = op_module.tab_inter(self.tab_info,
-                                      self.tab_data,
-                                      atype,
-                                      rij,
-                                      nlist,
-                                      natoms,
-                                      sw_lambda,
-                                      sel_a = self.sel_a,
-                                      sel_r = self.sel_r)
-            energy_diff = tab_atom_ener - tf.reshape(atom_ener, [-1, natoms[0]])
-            tab_atom_ener = tf.reshape(sw_lambda, [-1]) * tf.reshape(tab_atom_ener, [-1])
-            atom_ener = tf.reshape(inv_sw_lambda, [-1]) * atom_ener
-            energy_raw = tab_atom_ener + atom_ener
-        else :
-            energy_raw = atom_ener
 
-        energy_raw = tf.reshape(energy_raw, [-1, natoms[0]], name = 'o_atom_energy'+suffix)
-        energy = tf.reduce_sum(global_cvt_2_ener_float(energy_raw), axis=1, name='o_energy'+suffix)
-
-        net_deriv_tmp = tf.gradients (atom_ener, descrpt_reshape)
-        net_deriv = net_deriv_tmp[0]
-        net_deriv_reshape = tf.reshape (net_deriv, [-1, natoms[0] * self.ndescrpt])
-
-        force = op_module.prod_force_se_r (net_deriv_reshape,
-                                            descrpt_deriv,
-                                            nlist,
-                                            natoms)
-        if self.srtab is not None :
-            sw_force \
-                = op_module.soft_min_force(energy_diff, 
-                                           sw_deriv,
-                                           nlist, 
-                                           natoms,
-                                           n_a_sel = self.nnei_a,
-                                           n_r_sel = self.nnei_r)
-            force = force + sw_force + tab_force
-
-        force = tf.reshape (force, [-1, 3 * natoms[1]], name = "o_force"+suffix)
-
+    def prod_force_virial(self, atom_ener, natoms) :
+        [net_deriv] = tf.gradients (atom_ener, self.descrpt)
+        net_deriv_reshape = tf.reshape (net_deriv, [-1, natoms[0] * self.ndescrpt])        
+        force \
+            = op_module.prod_force_se_r (net_deriv_reshape,
+                                         self.descrpt_deriv,
+                                         self.nlist,
+                                         natoms)
         virial, atom_virial \
             = op_module.prod_virial_se_r (net_deriv_reshape,
-                                           descrpt_deriv,
-                                           rij,
-                                           nlist,
-                                           natoms)
-        if self.srtab is not None :
-            sw_virial, sw_atom_virial \
-                = op_module.soft_min_virial (energy_diff,
-                                             sw_deriv,
-                                             rij,
-                                             nlist,
-                                             natoms,
-                                             n_a_sel = self.nnei_a,
-                                             n_r_sel = self.nnei_r)
-            atom_virial = atom_virial + sw_atom_virial + tab_atom_virial
-            virial = virial + sw_virial \
-                     + tf.reduce_sum(tf.reshape(tab_atom_virial, [-1, natoms[1], 9]), axis = 1)
+                                          self.descrpt_deriv,
+                                          self.rij,
+                                          self.nlist,
+                                          natoms)
+        return force, virial, atom_virial
+    
 
-        virial = tf.reshape (virial, [-1, 9], name = "o_virial"+suffix)
-        atom_virial = tf.reshape (atom_virial, [-1, 9 * natoms[1]], name = "o_atom_virial"+suffix)
-
-        return energy, force, virial, energy_raw, atom_virial
-
-
-    def build_atom_net (self, 
-                        inputs,
-                        fparam,
-                        natoms,
-                        bias_atom_e = None,
-                        reuse = None,
-                        suffix = '') :
+    def _pass_filter(self, 
+                     inputs,
+                     natoms,
+                     reuse = None,
+                     suffix = '') :
         start_index = 0
         inputs = tf.reshape(inputs, [-1, self.ndescrpt * natoms[0]])
         shape = inputs.get_shape().as_list()
-        if bias_atom_e is not None :
-            assert(len(bias_atom_e) == self.ntypes)
-
+        output = []
         for type_i in range(self.ntypes):
-            # cut-out inputs
             inputs_i = tf.slice (inputs,
                                  [ 0, start_index*      self.ndescrpt],
                                  [-1, natoms[2+type_i]* self.ndescrpt] )
             inputs_i = tf.reshape(inputs_i, [-1, self.ndescrpt])
+            layer = self._filter_r(inputs_i, name='filter_type_'+str(type_i)+suffix, natoms=natoms, reuse=reuse, seed = self.seed)
+            layer = tf.reshape(layer, [-1, natoms[2+type_i] * self.get_dim_out()])
+            output.append(layer)
             start_index += natoms[2+type_i]
-            if bias_atom_e is None :
-                type_bias_ae = 0.0
-            else :
-                type_bias_ae = bias_atom_e[type_i]
-
-            # compute atom energy
-            layer = self._filter_r(inputs_i, name='filter_r_type_'+str(type_i)+suffix, natoms=natoms, reuse=reuse, seed = self.seed)
-            if self.numb_fparam > 0 :
-                ext_fparam = tf.reshape(fparam, [-1, self.numb_fparam])
-                ext_fparam = tf.tile(ext_fparam, [1, natoms[0]])
-                ext_fparam = tf.reshape(ext_fparam, [-1, self.numb_fparam])
-                layer = tf.concat([layer, ext_fparam], axis = 1)
-            for ii in range(0,len(self.n_neuron)) :
-                if ii >= 1 and self.n_neuron[ii] == self.n_neuron[ii-1] :
-                    layer+= self.one_layer(layer, self.n_neuron[ii], name='layer_'+str(ii)+'_type_'+str(type_i)+suffix, reuse=reuse, seed = self.seed, use_timestep = self.resnet_dt)
-                else :
-                    layer = self.one_layer(layer, self.n_neuron[ii], name='layer_'+str(ii)+'_type_'+str(type_i)+suffix, reuse=reuse, seed = self.seed)
-            final_layer = self.one_layer(layer, 1, activation_fn = None, bavg = type_bias_ae, name='final_layer_type_'+str(type_i)+suffix, reuse=reuse, seed = self.seed)
-            final_layer = tf.reshape(final_layer, [-1, natoms[2+type_i]])
-            # final_layer = tf.cond (tf.equal(natoms[2+type_i], 0), lambda: tf.zeros((0, 0), dtype=global_tf_float_precision), lambda : tf.reshape(final_layer, [-1, natoms[2+type_i]]))
-
-            # concat the results
-            if type_i == 0:
-                outs = final_layer
-            else:
-                outs = tf.concat([outs, final_layer], axis = 1)
-
-        return tf.reshape(outs, [-1])
-
+        output = tf.concat(output, axis = 1)
+        return output
 
     def _compute_dstats_sys_se_r (self,
                                   data_coord, 
@@ -452,4 +304,3 @@ class ModelSeR (Model):
             result = tf.reduce_mean(xyz_scatter, axis = 1) * res_rescale
 
         return result
-
